# aulas/aula 06/doc/openapi.yaml
openapi: 3.0.3
info:
  title: Projeto Controle de Filmes - SENAI Jandira
  description: |-
   API destinada ao controle de filmes para o projeto do senai
  termsOfService: http://swagger.io/terms/
  contact:
    email: kauan0alvess@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.4.26
...
servers:
  - url: http://
paths:
...
  /v1/senai/locadora/ator:
    post:
      tags:
        - Ator
      summary: Adiciona um novo ator
      description: Endpoint para enviar os dados do ator a serem cadastrados
      requestBody:
        description: JSON contendo os dados do novo ator a ser inserida no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ator'
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content-Type
        '500':
          description: Error Internal Server (Controller / Model)
    get:
      tags:
        - Ator
      summary: Lista todos os atores
      description: Listagem de todos os atores registradas no sistema
      responses:
        '200':
          description: Success Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AtorGet'              
        '404':
          description: Error Not Found
        '500':
          description: Error Interal Server (Controller / Model)
  /v1/senai/locadora/ator/{id}:
    get:
        tags:
          - Ator
        summary: Busca um ator pelo ID
        description: Retorna um ator pelo ID
        parameters:
          - name: id
            in: path
            description: ID para realizar a busca do ator
            required: true
            schema:
              type: integer
              format: int64
        responses:
          '200':
            description: Success Response
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/AtorGet'          
          '400':
            description: Error Bad Request
          '404':
            description: Error Not Found
          '500':
            description: Error Internal Server (Controller / Model)
    put:
      tags:
        - Ator
      summary: Atualizar um ator pelo ID
      description: Buscar e atualizar um ator pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca do ator
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar um ator
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ator'
        required: true
      responses:
        '200':
          description: Success Updated Item
        '400':
          description: Error Bad Request
        '404':
          description: Error Not Found
        '415':
          description: Error Content-Type
        '500':
          description: Error Internal Server (Controller / Model)
    delete:
      tags:
        - Ator
      summary: Excluir um ator pelo ID
      description: Buscar e excluir um ator pelo ID
      parameters:
        - name: id
          in: path
          description: ID para buscar o ator
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success Deleted Item
        '400':
          description: Error Bad Request 
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller / Model)

  # ENDPOINTS DA TABELA - CLASSIFICAÇÃO
  /v1/senai/locadora/classificacao:
    post:
      tags:
        - Classificacao
      summary: Adiciona uma nova classificação indicativa
      description: Endpoint para enviar os dados da classificação indicativa a serem cadastrados
      requestBody:
        description: JSON contendo os dados da nova classificação indicativa a ser inserida no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Classificacao'
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content-Type
        '500':
          description: Error Internal Server (Controller / Model)
    get:
      tags:
        - Classificacao
      summary: Lista todas as classificações indicativas
      description: Listagem de todas as classificações indicativas registradas no sistema
      responses:
        '200':
          description: Success Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Classificacao'              
        '404':
          description: Error Not Found
        '500':
          description: Error Interal Server (Controller / Model)
  /v1/senai/locadora/classificacao/{id}:
    get:
        tags:
          - Classificacao
        summary: Busca uma classificação indicativa pelo ID
        description: Retorna uma classificação indicativa pelo ID
        parameters:
          - name: id
            in: path
            description: ID para realizar a busca da classificação indicativa
            required: true
            schema:
              type: integer
              format: int64
        responses:
          '200':
            description: Success Response
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Classificacao'          
          '400':
            description: Error Bad Request
          '404':
            description: Error Not Found
          '500':
            description: Error Internal Server (Controller / Model)
    put:
      tags:
        - Classificacao
      summary: Atualizar uma classificação indicativa pelo ID
      description: Buscar e atualizar uma classificação indicativa pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca da classificação indicativa
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar uma classificação indicativa
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Classificacao'
        required: true
      responses:
        '200':
          description: Success Updated Item
        '400':
          description: Error Bad Request
        '404':
          description: Error Not Found
        '415':
          description: Error Content-Type
        '500':
          description: Error Internal Server (Controller / Model)
    delete:
      tags:
        - Classificacao
      summary: Excluir uma classificação indicativa pelo ID
      description: Buscar e excluir uma classificação indicativa pelo ID
      parameters:
        - name: id
          in: path
          description: ID para buscar a classificação indicativa
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success Deleted Item
        '400':
          description: Error Bad Request 
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller / Model)

  # ENDPOINTS DA TABELA - DIRETOR
  /v1/senai/locadora/diretor:
    post:
      tags:
        - Diretor
      summary: Adiciona um novo diretor
      description: Endpoint para enviar os dados do diretor a serem cadastrados
      requestBody:
        description: JSON contendo os dados do novo diretor a ser inserida no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Diretor'
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content-Type
        '500':
          description: Error Internal Server (Controller / Model)
    get:
      tags:
        - Diretor
      summary: Lista todos os diretores
      description: Listagem de todos os diretores registradas no sistema
      responses:
        '200':
          description: Success Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DiretorGet'              
        '404':
          description: Error Not Found
        '500':
          description: Error Interal Server (Controller / Model)
  /v1/senai/locadora/diretor/{id}:
    get:
        tags:
          - Diretor
        summary: Busca um diretor pelo ID
        description: Retorna um diretor pelo ID
        parameters:
          - name: id
            in: path
            description: ID para realizar a busca do diretor
            required: true
            schema:
              type: integer
              format: int64
        responses:
          '200':
            description: Success Response
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/DiretorGet'          
          '400':
            description: Error Bad Request
          '404':
            description: Error Not Found
          '500':
            description: Error Internal Server (Controller / Model)
    put:
      tags:
        - Diretor
      summary: Atualizar um diretor pelo ID
      description: Buscar e atualizar um diretor pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca do diretor
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar um diretor
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Diretor'
        required: true
      responses:
        '200':
          description: Success Updated Item
        '400':
          description: Error Bad Request
        '404':
          description: Error Not Found
        '415':
          description: Error Content-Type
        '500':
          description: Error Internal Server (Controller / Model)
    delete:
      tags:
        - Diretor
      summary: Excluir um diretor pelo ID
      description: Buscar e excluir um diretor pelo ID
      parameters:
        - name: id
          in: path
          description: ID para buscar o diretor
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success Deleted Item
        '400':
          description: Error Bad Request 
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller / Model)

  # ENDPOINTS DA TABELA - FILME
  /v1/senai/locadora/filme:
    post:
      tags:
        - Filme
      summary: Adiciona um novo filme
      description: Endpoint para enviar os dados do filme a serem cadastrados
      requestBody:
        description: JSON contendo os dados do novo filme a ser inserido no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content-Type
        '500':
          description: Error Internal Server (Controller / Model)
    get:
      tags:
        - Filme
      summary: Lista todos os filmes
      description: Listagem de todos os filmes registrados no sistema
      responses:
        '200':
          description: Success Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FilmeGet'          
        '404':
          description: Error Not Found
        '500':
          description: Error Interal Server (Controller / Model)
  /v1/senai/locadora/filme/{id}:
    get:
      tags:
        - Filme
      summary: Busca um filme pelo ID
      description: Retorna um filme pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca do filme
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FilmeGet'          
        '400':
          description: Error Bad Request
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller / Model)
    put:
      tags:
        - Filme
      summary: Atualizar um filme pelo ID
      description: Buscar e atualizar um filme pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca do filme
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar um filme
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
        required: true
      responses:
        '200':
          description: Success Updated Item
        '400':
          description: Error Bad Request
        '404':
          description: Error Not Found
        '415':
          description: Error Content-Type
        '500':
          description: Error Internal Server (Controller / Model)
    delete:
      tags:
        - Filme
      summary: Excluir um filme pelo ID
      description: Buscar e excluir um filme pelo ID
      parameters:
        - name: id
          in: path
          description: ID para buscar o filme
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success Deleted Item
        '400':
          description: Error Bad Request
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller / Model)

  # ENDPOINTS DA TABELA - FOTO
...
components:
  schemas:
...
    Ator:
      required:
      - id
      - nome
      - data_nascimento
      - inicio_carreira
      - biografia
      - id_sexo
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
          example: Caio
          description: O número máximo de caracteres é 100.
        data_nascimento:
          type: string
          format: date
          example: "2000-01-01"
        inicio_carreira:
          type: string
          pattern: '^\d{4}$'
          example: '2020'
          description: Ano no formato AAAA (máximo de 4 dígitos).
        biografia:
          type: string
          description: O número máximo de caracteres é ilimitado.
        id_sexo:
          type: integer
        atividade:
          type: object
          example: {"id": 1}
          description: Coloque , depois de abrir e fechar o json.
        foto:
          type: object
          example: {"id": 1}
          description: Coloque , depois de abrir e fechar o json.
        nacionalidade:
          type: object
          example: {"id": 1}
          description: Coloque , depois de abrir e fechar o json.
    Classificacao:
      required:
      - id
      - classificacao
      type: object
      properties:
        id:
          type: integer
        classificacao:
          type: string
          example: "+14"
          description: "O número máximo de caracteres é de 5, caso seja livre, guarde a letra L"
        descricao:
          type: string
          description: O número máximo de caracteres é ilimitado.
        idade_minima:
          type: integer
    Diretor:
      required:
      - id
      - nome
      - data_nascimento
      - inicio_carreira
      - id_sexo
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
          example: José
          description: O número máximo de caracteres é 100.
        data_nascimento:
          type: string
          format: date
          example: "2000-01-01"
        inicio_carreira:
          type: string
          pattern: '^\d{4}$'
          example: '2020'
          description: Ano no formato AAAA (máximo de 4 dígitos)
        id_sexo:
          type: integer
        atividade:
          type: object
          example: {"id": 1}
          description: Coloque , depois de abrir e fechar o json.
        foto:
          type: object
          example: {"id": 1}
          description: Coloque , depois de abrir e fechar o json.
        nacionalidade:
          type: object
          example: {"id": 1}
          description: Coloque , depois de abrir e fechar o json.
    Filme:
      required:
        - id
        - nome
        - sinopse
        - capa
        - data_lancamento
        - duracao
        - id_classificacao
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
          description: O número máximo de caracteres é 80.
        sinopse:
          type: string
          description: O número máximo de caracteres é ilimitado.
        capa:
          type: string
          description: O número máximo de caracteres é 255.
        data_lancamento:
          type: string
          format: date
          example: "2026-01-01"
        duracao:
          type: string
          example: "02:30"
          pattern: '^(?:[01]\d|2[0-3]):[0-5]\d$'
        valor:
          type: number
          format: double
          example: "999.99"
          description: O número máximo de caracteres é de 9 centenas, 9 dezenas e 9 unidades (999) e 2 casas decimais (.99).
        avaliacao:
          type: number
          format: double
          example: "5.0"
          description: O número máximo de caracteres é de 5 unidade (5) e 1 casa decimal (.9), porém tendo como valor máximo 5.0.
        id_classificacao:
          type: integer
        ator:
          type: object
          example: {"id": 1}
          description: Coloque , depois de abrir e fechar o json.
        diretor:
          type: object
          example: {"id": 1}
          description: Coloque , depois de abrir e fechar o json.
        genero:
          type: object
          example: {"id": 1}
          description: Coloque , depois de abrir e fechar o json.
...
    AtividadeGet:
      required:
      - id
      - area_atuacao
      type: object
      properties:
        id:
          type: integer
        area_atuacao:
          type: string
          description: O número máximo de caracteres é 100.
        ator:
          type: array
          items:
            $ref: '#/components/schemas/AtorAtividadeGet'
        diretor:
          type: array
          items:
            $ref: '#/components/schemas/DiretorAtividadeGet'
    AtorAtividadeGet:
      required:
      - id
      - nome
      - data_nascimento
      - inicio_carreira
      - biografia
      - id_sexo
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
          example: Caio
        data_nascimento:
          type: string
          format: date
          example: "2000-01-01"
        inicio_carreira:
          type: integer
          example: 2020
        biografia:
          type: string
        id_sexo:
          type: integer
    AtividadeAtorGet:
      type: object
      properties:
        id:
          type: integer
        area_atuacao:
          type: string
        diretor:
          type: array
          items:
            $ref: '#/components/schemas/DiretorAtividadeGet'
        ator:
          type: array
          items:
            $ref: '#/components/schemas/AtorAtividadeGet'
    DiretorAtividadeGet:
      required:
      - id
      - nome
      - data_nascimento
      - inicio_carreira
      - id_sexo
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
          example: Caio
        data_nascimento:
          type: string
          format: date-time
          example: "2000-01-01"
        inicio_carreira:
          type: integer
          example: 2020
        id_sexo:
          type: integer
    AtorGet:
      required:
      - id
      - nome
      - data_nascimento
      - inicio_carreira
      - biografia
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        data_nascimento:
          type: string
          format: date-time
        inicio_carreira:
          type: integer
        biografia:
          type: string
        sexo:
          type: array
          items:
            $ref: '#/components/schemas/SexoAtorGet'
        nacionalidade:
          type: array
          items:
            $ref: '#/components/schemas/NacionalidadeAtorGet'
        foto:
          type: array
          items:
            $ref: '#/components/schemas/FotoAtorGet'
        atividade:
          type: array
          items:
            $ref: '#/components/schemas/AtividadeAtorGet'
    SexoAtorGet:
      type: object
      properties:
        id:
          type: integer
        sigla:
          type: string
        sexo:
          type: string
    FotoAtorGet:
      type: object
      properties:
        id:
          type: integer
        foto:
          type: string
    NacionalidadeAtorGet:
      type: object
      properties:
        id:
          type: integer
        nacionalidade:
          type: string
...
    DiretorGet:
      required:
      - id
      - nome
      - data_nascimento
      - inicio_carreira
      - sexo
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
          example: Caio
          description: O número máximo de caracteres é 100.
        data_nascimento:
          type: string
          format: date-time
          example: "2000-01-01"
        inicio_carreira:
          type: integer
          example: 2020
          description: Ano no formato AAAA (máximo de 4 dígitos).
        sexo:
          type: array
          items:
            $ref: '#/components/schemas/SexoGet'
        nacionalidade:
          type: array
          items:
            $ref: '#/components/schemas/NacionalidadeGet'
        foto:
          type: array
          items:
            $ref: '#/components/schemas/FotoGet'
        atividade:
          type: array
          items:
            $ref: '#/components/schemas/AtividadeGet'
    FilmeGet:
      required:
      - id
      - nome
      - sinopse
      - capa
      - data_lancamento
      - duracao
      - valor
      - avaliacao
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
          description: O número máximo de caracteres é 100.
        sinopse:
          type: string
          description: O número máximo de caracteres é ilimitado.
        capa:
          type: string
          description: URL da capa do filme.
        data_lancamento:
          type: string
          format: date-time
          example: "2000-01-01T00:00:00.000Z"
        duracao:
          type: string
          example: "01:20:00"
        valor:
          type: string
          example: "10.00"
        avaliacao:
          type: string
          example: "5.00"
        classificacao:
          type: array
          items:
            $ref: '#/components/schemas/ClassificacaoFilmeGet'
        genero:
          type: array
          items:
            $ref: '#/components/schemas/GeneroFilmeGet'
        diretor:
          type: array
          items:
            $ref: '#/components/schemas/DiretorFilmeGet'
        ator:
          type: array
          items:
            $ref: '#/components/schemas/AtorFilmeGet'
    ClassificacaoFilmeGet:
      required:
      - id
      - classificacao
      - descricao
      - idade_minima
      type: object
      properties:
        id:
          type: integer
        classificacao:
          type: string
          example: "L"
          description: O número máximo de caracteres é 10.
        descricao:
          type: string
          description: O número máximo de caracteres é 255.
        idade_minima:
          type: integer
          example: 0
    GeneroFilmeGet:
      required:
      - id
      - genero
      type: object
      properties:
        id:
          type: integer
        genero:
          type: string
          description: O número máximo de caracteres é 100.
    DiretorFilmeGet:
      required:
      - id
      - nome
      - data_nascimento
      - inicio_carreira
      - id_sexo
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
          example: Caio
          description: O número máximo de caracteres é 100.
        data_nascimento:
          type: string
          format: date-time
          example: "2000-01-01T00:00:00.000Z"
        inicio_carreira:
          type: integer
          example: 2020
          description: Ano no formato AAAA (máximo de 4 dígitos).
        id_sexo:
          type: integer
    AtorFilmeGet:
      required:
      - id
      - nome
      - data_nascimento
      - inicio_carreira
      - biografia
      - id_sexo
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
          example: Caio
          description: O número máximo de caracteres é 100.
        data_nascimento:
          type: string
          format: date-time
          example: "2000-01-01T00:00:00.000Z"
        inicio_carreira:
          type: integer
          example: 2020
          description: Ano no formato AAAA (máximo de 4 dígitos).
        biografia:
          type: string
        id_sexo:
          type: integer
    FotoGet:
      required:
      - id
      - foto
      type: object
      properties:
        id:
          type: integer
        foto:
          type: string
          description: O número máximo de caracteres é de 255.
...
    NacionalidadeGet:
      required:
      - id
      - nacionalidade
      type: object
      properties:
        id:
          type: integer
        nacionalidade:
          type: string
          description: O número máximo de caracteres é 100.
        diretor:
          type: array
          items:
            $ref: '#/components/schemas/DiretorNacionalidadeGet'
        ator:
          type: array
          items:
            $ref: '#/components/schemas/AtorNacionalidadeGet'
...
    SexoGet:
      required:
      - id
      - sigla
      - sexo
      type: object
      properties:
        id:
          type: integer
        sigla:
          type: string
          example: "M"
          description: O número máximo de caracteres é de 3.
        sexo:
          type: string
          example: "Masculino"
          description: O número máximo de caracteres é de 15.
  
  # CAMINHOS PARA DOCUMENTAR
...